File System Replication
FileSystem
==========

This project implements a file system with
clients trying to write objects to servers. A client tries to write any
object on 3 servers and an object is eligible to be written only when
atleast 2 out of the 3 servers are accessible. The project also takes
care of total ordering while writing objects so that in a scenario when
multiple clients try to write the same object on same servers, the order
in which the object is written in each of the server is consistent.